regress: fetch -b/got.conf/wt/remote HEAD interaction
The added test checks we honour the correct precedence of branches to be
fetched, which is determined by: the -b flag; branches set in got.conf;
the work tree's current branch; and the remote repository's HEAD branch.

If a branch is specified with -b, only fetch this branch and if it does
not exist, simply report to the user--do not fallback to other branches.
Otherwise, if branches are either not set in got.conf or are set but do
not exist on the server:
  - if invoked in a work tree, fallback to the work tree's current branch
  - if not in a work tree or the work tree's current branch does not exist
    on the server:
      - fallback to remote HEAD.

With this, only the -b case can result in an unsuccessful fetch due to the
requested branch(es) not existing on the server. All other cases will either
fetch updates or report "Already up-to-date" (excepting errors).

ok stsp@