fix GOT_IMSG_COMMIT_TRAVERSAL_REQUEST
The sending and receiving side are sending the data differently.  It
works now by chance since a struct got_object_id is "just" a
SHA1_DIGEST_LENGTH sized buffer, but will break in the future.
Furthermore, the structure of the data as described in got_lib_privsep.h
is not respected, as the path_len field is not transmitted.

Change it to send/receive a whole struct got_object_id, the path without
the NUL and the proper length.  (The path is also actually mandatory, so
assert it too.)

improvements + ok stsp@