portable: configure: use config.h
So that the command-line flags to CC don't get drowned out with copious
-D flags, move this detection to config.h instead.

Note that to reduce the code churn, the existing include/got_compat.h
header file has been renamed, and the generated file replaces it, but
with the contents of the original header file included.

Eventually, we might want to move some od the logic in
include/got_compat2.h into AH_VERBATIM() blocks, but for now, this is a
good compromise.