portable: configure: split out dependencies
Rather than assume all dependencies are required for all programs, split
them out.

This will make packaging easier, as well as splitting the code to use
subprojects.

Note that due to the use of config.h semantics, in most cases the
got_compat.h header file is now at the top of the .c file it is included
in, so that it can handle the system header inclusion properly.