gotd requires a config file; don't fail silently when gotd.conf cannot be read
Fixes an issue encountered by mlarkin@ where the auth process was working
with an empty list of repositories and kept saying it could not find a
repository which was obviously listed in gotd.conf. Now we can see errors
from fopen() in the logs instead. Old code was copied from gotwebd where
the config file is optional.