gotwebd: handle short reads and timeouts
If a short read happens, or if all the fastcgi record to read don't fit
in the buffer, gotwebd fails to continue reading and hits the timeout.
If this happens before gotweb_process_request is called, it will crash
in gotweb_free_transport since c->t will be NULL.

This register the event with EV_PERSIST so fcgi_request is called again
when there's more to read and guards gotweb_free_transport.  It also
makes spaces for the records as soon as they're successfully parsed.

With lots of help from stsp

ok stsp@