tog: resize log view if toggling fullscreen from child view
We call resize_log_view() when toggling 'F'ullscreen from the log view, but
forgot to do this when toggling fullscreen from a child view if its parent
is the log view. This fixes a bug that prevents scrolling down the log view
from the diff view with the J keymap if already on the last loaded commit.
Regress test in the following commit.

ok stsp@