simplify lib/send.c reference handling
Reorganize the reference validation and pathlist generation by
removing the reflist and building a pathlist directly.  The pathlist
entries record the object id in their extra data pointer, which also
allows several redundant reference lookups to be skipped.

This will eventually simplify sending target reference names that do
not match the local repo by adding another parameter to
insert_sendable_ref for a remote reference.  This remote name will be
added to the pathlist, but validation and object id lookups will
continue to be performed with the local reference.

ok jamsek