make 'gotadmin load' always read data from standard input
This provides better symmetry with 'gotadmin dump', and allows us to pass
the list of references as command line arguments, replacing the -b option.

The -l option now takes an argument which specifies the bundle file rather
than requiring data to be fed on stdin with -l which feels awkward.

ok op@