got: improve reporting accuracy in branch listing output
As suggested by stsp, show the out-of-date symbol for the work tree branch if
it is not only out of date in relation to the branch tip but also if it
contains mixed commits. Update regress and add new test to check for this case.
Also, explain the mixed commit condition of * and ~ markers denoting work
tree state in got(1) branch -l docs.

ok stsp@