tog: don't show the base commit marker in limit view
The limit commit queue is invariably filled with different commits to the
queue containing all commits, so the index of the work tree's base commit in
the real queue corresponds to a different commit in the limit queue. As such,
the marker is drawn on an incorrect commit. Rather than fix this to draw the
marker on the correct commit if it happens to be in the limit queue, don't
draw the marker at all in limit view. As pointed out by Mikhail on the list,
this information is not wanted in the limit view. Patch and report by Mikhail.