prevent a double-free in got_worktree_commit
If creating the /tmp display diff for a commitable item failed we
would free the commitable item while it was already on the path list.
Later on when the path list was freed in got_worktree_commit() a
double-free would be detected and the program would be aborted.

Found by gonzalo@
ok op@