fix deletion of missing locally-added files
If a locally added file in A status gets deleted from disk it will
move into missing ! status. If the user then decides to delete the
file we must remove the file index entry immediately rather than
flagging the file for removal during the next commit operation.
The file was never committed and lacks a base-blob and base-commit,
so it cannot be removed during the next commit. As a result the
commit operation was reporting a bogus "out of date" error.

Found while diagnosing a related bug reported by gonzalo@
ok op@