simplify usage of the 'mesg' histedit script command
The 'mesg' script command now requires a commit ID as its argument,
rather than being tied to a preceding 'pick' or 'edit' command.

The old model was too confusing for new users, in particular for people
used to Git's rebase -i squash semantics.
The 'mesg' command is now semantically equivalent to the 'pick' command
and additionally opens the log message in an editor.

The new syntax is simpler to use but also requires that we drop support
for one-line log messages inside the histedit script, with a commit ID
taking its place in the argument space. We don't believe this feature was
used much, and that a simplified usage model is more beneficial overall.

Patch by Lorenz (xha)
ok jamsek