gotwebd: get rid of proc.[ch]
proc.c really shines when there's a network of different types of
processes, potentially with a various number of instances each,
that needs to exchange messages.  Gotwebd instead has a much simpler
design, and using proc.c causes more overhead (/headaches) than it
solves.

So, this attempts to provide the same functionalities but with a
much simpler implementation that fits gotwebd better.

ok stsp@