gotwebd: try to improve copy-paste from BLOB pages
This moves the space inside the line anchor and makes it non-selectable,
so that when copy-pasting, only the content of line lines is actually
copied.

At the moment, this seems to behave slightly incorrectly in firefox
which inserts a double line break due to the presence of unselectable
elements inside the selection.  Hopefully it will be fixed, chromium
behaves fine.

It also still works decently on text-browsers.

ok stsp@