add an xfail test for a case where rebase fails to forward a branch
Because 'got rebase' only does a first-parent traversal it will try to
rebase commits which appear in the history of a branch, even when the
branch to be rebased is already based on that history. This results in
spurious merge conflicts as existing changes get re-applied.

The desired behaviour would be that 'got rebase' forwards the branch,
as it does when the 'got merge -M' command used by this test case is
replaced by a simple 'got merge' which avoids creating a merge commit.

Problem reported by naddy in the "Landry's firefox repository" thread:
https://marc.gameoftrees.org/mail/1706721001.20565_0.html