got-notify-http: change the layout of the json
Split the author/committer in sub-fields, rename author -> committer
in the short format parser and cache the author for the long format
in case it's the same user as the committer.  Put a copy of the
first line of the commit message in the "short_message" field.

There's some redundancy now, but it's to make consuming easier.

ok stsp@