got-notify-http: refactor jsonify()
To support the other types of notifications, we can't just peek at
the first byte and decide wether it's a short or long format.
Refactor the parsing so that there's one main entrypoint, jsonify(),
that calls jsonify_comment() or jsonify_comment_short() depending
on the format of the line.  Other "line types" will be added as a
follow-up to support the other notification types.