got-notify-http: fix I/O with TLS
tls_write() may return TLS_WANT_READ.  In that case, we're stuck
trying to read while we have the request still in the buffer.  So,
we have to retry bufio_write() regardless of the POLLIN/POLLOUT
state.

We'd need this only in the TLS case, but in practice it doesn't
harm for the plaintext case too.  In fact, we're either waiting to
flush the request or we're only reading the server reply.

ok stsp