fix diff view J keymap to diff the next commit in the log
When the last visible commit in the log view is selected while in a log/diff
horizontal split, and fullscreen is then toggled in the diff view, the user
cannot scroll to the next commit with the diff view J keymap due to a bad
predicate guarding a clamp to prevent scrolling off-screen that should only
be applied when in the horizontally split log view and not when scrolling
from the diff view. The fix also reveals an off-by-one that's only applicable
in the hsplit case in the subsequent commit loading logic that also breaks J
when attempting to scroll beyond the last loaded commit.
New regress added to cover these cases.

ok op@ (caveated that he's unsure exactly why the OB1 fix works)