gotwebd: change how we deal with a non-existant config file
We don't need a config file since we have some defaults, but we
also need a way to set them.  After the change to move `listen on'
to a global rule, we don't add the default socket anymore when
there's no config file.  Instead, change how we deal with an
*existing* config file and leave the rest of parse_function() to
handle the case of missing knobs to set up.