prepare got_object_parse_tree for sha256
Since got_object_parse_tree_entry is a function that has to be as fast
as possible, the idea to handle both types of hashes is to keep track of
the hashing algorithm used and the length of the digest, and then
eventually adding a "gap" when sending in privsep.c so that the
receiving side can always read a fixed-size buffer.  This is to avoid
doing memcpy() which can be costly in the hot path.

ok stsp@