swap trim_obj_id behaviour
trim_obj_ids is a bit weird in that you have to specify how many
characters to drop *from the end* instead of how many to keep from the
beginning.  This comes in the way of enabling some parts of the regress
for sha256 since object IDs are longer, and a `trim_obj_id 38` will not
yield 2 characters, but 26!

Instead, use the numeric argument to mean the number of characters to
keep from the beginning.  Then `trim_obj_id 2` will always return the
first two characters, regardless of the length of the hash.

ok stsp@