fix diff3 mishandling of malformed text files
If histedit, rebase, or merge involves a malformed text file (i.e., one
without a terminating newline), we generate an incorrect ed script as the
command gets appended to the final line rather than on a newline. This
produces an unreported fatal error and either truncates the malformed file
or appends some of the ed script to it. Make sure we handle such cases by
inserting the newline before the ed command.

reported and tested by naddy, style tweaks and ok stsp@