work around thin bundles in regress and add a TODO
git bundle hardcodes --thin when using an exclude base and doesn't
provide a way to disable it.  Somehow, currently, it decides to use
a thin pack only in sha256 and not in sha1.  For now however, change
that test to use gotadmin dump instead to generate an incremental
bundle.

We'll have to support in some ways the thin packs to be completely
compatible with git bundles, so add a TODO.

discussed with stsp.