fix signedness issue in got_pack_index() after sha256 work
Later we try to see to -digest_len from the end of the file, and
this on 32bit arches has some issues: size_t is 32bit and being
unsigned becomes a huge number that then is casted to a off_t
(signed, and larger) and makes us seek very, very far from the end
of file.  Then, read(2) returns zero because we're at end of file
and trying to send packs to gotd fails with "I/O error".  This was
introduced when SHA1_DIGEST_LENGTH was converted to digest_len.

ok jamsek, stsp