rename idlen to digest_{,string_}len
In my early sha256 work I've used `idlen' to parametrize the digest
length, but that's ambiguous since it could refer either to the
digest length (in binary form) or the digest _string_ length
(hexadecimal).

So, change the few offenders to either digest_len or digest_string_len,
which is a "naming scheme" I've already used in the rest of the
tree.

ok stsp@