got/cvg info: print work tree version
This exends got_worktree_path_info() to resemble the "transaction"-like
interface we have for rebase, patch etc.  There's a prepare() routine
that returns the fileindex, and locks the worktree, and a complete()
function to free the fileinedx and release the lock.  This way, we
can open only once the fileindex in cmd_info() and have the chance
to ask for its version.

ok stsp@