gotwebd: unbreak tags rendering
if qs->commit is set got_get_repo_tags interprets it as an index
for the tags, i.e. to skip tags until one with that id is found.
So we can't set qs->commit (intended to unbreak the README links)
before it, we have to do it afterwards.

ok stsp@