use _POSIX_HOST_NAME_MAX instead of HOST_NAME_MAX
this is to help -portable.  Some systems don't define HOST_NAME_MAX
at all, and assume instead that code will use sysctl(_SC_HOST_NAME_MAX)
with a VLA or a dynamic allocation, which is completely backward
and useless.  So, use the _POSIX hammer on this one too.

it's dirty, but it's better to strive to avoid gratious -portable
changes.

reported and discussed with Thomas Adam.