fix pack file creation in the presence of tagged tag objects
If a repository contains a tag that tags another tag we could fail
with a "bad object data" error while creating pack files because
the packed-object-enumeration code assumed that only commits get
tagged, trying to parse the tagged tag object as if it were a commit.

This issue affected 'got send' and 'gotadmin pack'.

There is probably more work to do here because other weird cases are
known to exist in the wild, such as git.git's refs/tags/junio-gpg-pub
which tags a blob.

Problem with 'got send' reported by jrick@