gotwebd: ensure child procs inherit non-default config
If a non-default config filepath is specified with -f, it is effectively
ignored as child processes always either parse the default config file
(/etc/gotwebd.conf) if it exists, or use defaults. Options specified in
the config file used by server procs (e.g., chroot, user) are ignored.
If -f is used with a non-default config file, pass it to child procs.

ok stsp@