gotwebd: improve performance of repository age calculations
Instead of sorting refs by name and then hunting for the youngest
commits in that list, obtain an appropriately sorted list via
got_ref_cmp_by_commit_timestamp_descending and then pick the first
ref from this list.

And in the case were we already know which ref to check, just open
this ref directly instead of listing all the refs.

Easier to read and should improve performance in both cases.

ok op, jamsek