implement tog work tree diff support via log view and CLI
Work tree diffs can now be viewed with `tog diff [-s] [path ...]` or by
selecting a work tree entry from the log view. If there are unstaged
changes, a work tree entry is displayed at the top of the log view;
likewise if there are staged changes. If there are both staged and
unstaged changes, two work tree entries are displayed. The log view can
be refreshed with ^l when new changes are staged, unstaged, or reverted
in the work tree, but each time a work tree entry is opened from the log
view, or an active work tree diff view is refreshed (e.g., A, [,], w) an
up-to-date diff is always generated using on disk changes.

ok stsp@