tog: remove author validation of log work tree entries
It's not the place of tog to validate such things. The author in this
case is for display purposes only. Got will alert the user in the event
the bad author value is set when it's needed to write to the repository.
Prompted by Lucas's report of tog's fatal error when unable to derive
an author to display in log view entries of a dirty work tree.

ok lucas@