tog: make missing author for worktree entries non-fatal
Reported by Lucas: if we can't find an author in the work tree or repo
got.conf, git's config file, or the GOT_AUTHOR environment variable, set
the author to the empty string for displaying worktree entries rather
than hit the user with a fatal error. Based on Lucas's patch: the only
change is to check for GOT_ERR_COMMIT_NO_AUTHOR before strdup(3)ing "".

ok lucas@ ("go ahead and commit" on IRC)