tog: make ref view selection of non-commit tags non-fatal
Tags can point to all git objects: commits, trees, blobs, and tags.
Selecting a tag that points to any object other than a commit causes a
fatal error. Instead, report a message to the status line. Similarly,
nested tags may resolve to a commit, which currently errors. Instead,
keep peeling till we reach the bottom and if it's a commit, use it for
the requested view.

ok stsp@