rework got-read-pack's commit coloring loop
Port the parent commit coloring fix to got-read-pack, and ensure that
it starts off with the same state as the main process.
got-read-pack did not have access to the main process's ids array, and
was thus working with a different initial state. With these changes the
same commit traversal happens regardless of whether coloring is "offloaded"
to got-read-pack or not (verified manually by placing debug printfs).

ok jamsek