fix diffstat path order bug in field width computation
The order in which get_diffstat() receives paths may not be the same as
the lexicographic order of the changed paths path list, so do not use
RB_MAX() to get the path that will be used in the current max path len
check. Use the path entry returned by the current got_pathlist_insert()
call instead. In such cases that get_diffstat() does not receive paths
in lexicographic order, this can produce a bogus diffstat max_path_len
if the last lexicographically sorted path entry doesn't have the longest
pathname of all changed paths in the diff, which manifests as misaligned
diffstat path separators as reported by stsp on IRC.

ok stsp@