introduce gotsys-sshdconfig to rewrite /etc/ssh/sshd_config and restart sshd
A gotsysd-managed system relies on sshd for network transport. We need to
let gotsysd take over the configuration of sshd to tweak a few settings:

Enable password-less login for the anonymous user, Disable all forwarding.
Restrict pty allocation to users in the wheel group since we expect nobody
else to be logging in via ssh for remote access.