potential fix for use-after-free in lib/repository.c's match_packed_object()
In a gotwebd core file on got.gameoftrees.org this loop crashed with a
pe pointer which had already been freed. The retries counter was still
zero, which means we passed the timestamp check when the loop iteration
began. Then after a call to got_object_get_type() the pack index list
had been recreated, rendering the pe pointer invalid. The code crashed
when RB_FOREACH tried to run the next iteration using this bad pointer.
The avoid this issue, move the timestamp check right below the call to
got_object_get_type(). This way, the 'goto retry' should take effect
earlier and restart the entire loop with a fresh 'pe' instead of crashing.