pass socket and connection limits to gotd listen process via imsg
This is a first step towards parsing gotd.conf only once at startup,
rather than parsing it whenever a new child process spawns. The current
behaviour is bad because gotd child processes will run with a different
configuration as soon as /etc/gotd.conf is edited, rather than requiring
a restart for configuration changes to take effect.
This commit only fixes this problem for the listen process. More to follow.