stop using got_repo_map_path() in gotwebd
got_repo_map_path() is supposed to figure out whether an input path
corresponds to an on-disk file in a non-bare git repository, and map
such paths to in-repository paths. In other cases it will just return
the input path as-is.

Since gotwebd only deals with repository-relative input paths in the
first place, it can just use input paths from query strings as-is.

Avoids realpath errors seen in my gotwebd server logs, such as:
  gotweb_render_blame: got_output_file_blame: realpath: template/tmpl.c: \
    Permission denied
With the blame page showing a header but no file content.

ok op@