portable: add a release script
This script automates the creation of the dist tarball, including
verifying the tarball by unpacking and compiling it.

Note that, the CHANGELOG should have been filled out already prior to
running this script, so that it's included in the tarball; but that
commit will need amending with the updated changes this script creates.

Also assumes ed(1) is installed which is slightly more portable (and
available) across different *nix systems, unlike sed(1)'s -i flag.