tweak 'got status' and 'got add' ignores handling for better git compatibility
Skip past leading slashes in ignore patterns. In gitignore(7), these
trigger the behavior got has by default. Add regress test.

Lift documentation of ignore patterns to a subsection, referenced by
the `add` and `status` docs.

Refine documentation of ignore patterns to discuss the extensions to
glob(7) as a whole, and then list caveats relative to git(1) and
cvs(1).

Add a file to the test worktree in the main gitignore regress test in
order to register a known delta to gitignore(7) that was ambiguously
documented previously: for the pattern a/**/foo, git will ignore
a/foo, but got will not.

Patch by: Evan Silberman