untangle committer and author_time parameters of got_worktree_commit()
Make the semantics of the author_time parameter self-contained by
treating author_time == 0 as the fallback case, instead of treating
committer == NULL as the fallback case. Having the semantics of one
parameter depend on the semantics of another might lead to subtle
bugs introduced during potential future refactoring of this code.