stop trying to start gotd from gotsys-apply-conf if gotd is not running
We should assume that gotd is always running, and we should not be
automatically restarting gotd if it has crashed.

This misfeature was added early to make setup of the gotsysd regression
test suite easier. It was only half working since the attempt to connect
to gotd's unix socket after starting it failed due to pledge violations.