fix gotd branch protection rejecting commits that already exist on server
When a reference update sends an empty pack file because the objects
already exist on a different branch, gotd was wrongly rejecting the
reference update. The logic which protects references from moving
kicked in, but didn't account for this particular case. Run a YCA
check to ensure the client is only appending new commits to the
branch and allow the ref update to proceed if so. Add test coverage.

Problem reported by martijn@