fix gotd reload when /etc/gotd-secrets.conf is used
When sending the configuration to the newly reloaded gotd process we must
send the path to the secrets file if it is being used. Otherwise, the new
process will die with "unexpected amount of data received in imsg" errors.

Test coverage will be added later as part of some new gotsysd tests
which I have in progress.