fix off-by-one during deltification in maximum stretch size calculation
Fix an off-by-one which causes invalid deltas to be written when common
file sections exceed the maximum size which can be represented in a
base-copy delta operation.

This bug causes an invalid pack file to be written which neither gotadmin
nor git will index successfully.

Add a test which triggers the problem:

got-index-pack: delta application result size mismatch: \
	actual: 65536 expected: 16777216: bad delta

ok op@