regress: replace "sed -i" with ed(1) for portable in-place editing
"sed -i" is fundamentally unportable.  GNU and OpenBSD sed(1) treat
the extension for the backup file as an optional argument and use
"sed -i" for no backup file.  FreeBSD sed(1) treats the extension
as an obligatory argument and uses "sed -i ''" for no backup file.
There is no single syntax that works for both.