gotwebd: avoid the pipe; forward the socket directly
Instead of keeping a pipe between the sockets and the gotwebd
process, just forward the socket directly.  This needs a little bit
of care because once we enter process_request() the socket (possibly)
no longer there, so on the socket side use the client_status as a
flag to signal the caller that we're done.  I hope to improve this
in follow-ups.

The motivation behind this is to avoid a (possible, but quite easily
triggerable) race in the current code where the other side might
process GOTWEBD_IMSG_REQ_DONE before draining the pipe, truncating
the output page.

ok stsp@